+++ public/times/delete.php
@@ -1,6 +1,12 @@
 <?php 
 include('../../includes/db.php'); 
 
+session_start();
+if (!isset($_SESSION["user_pk"])) {
+    header("Location: ../login.php");
+    exit;
+}
+
 $id = $_GET['id'];
 
 $sql = "DELETE FROM times WHERE id=$id";
